Added casting for GCC update.

+++ Basic/main.cpp
@@ -30,7 +30,7 @@ static void printTask(void *args) {
   struct timeval now;
   gettimeofday(&now, NULL);
   printf(
-    "Now: %u-%u-%u %s %02u:%02u:%02u (%ld.%06ld)\n",
+    "Now: %u-%u-%u %s %02u:%02u:%02u (%lu.%06ld)\n",
     t.tm_year + 1900,
     t.tm_mon + 1,
     t.tm_mday,
@@ -38,7 +38,7 @@ static void printTask(void *args) {
     t.tm_hour,
     t.tm_min,
     t.tm_sec,
-    now.tv_sec,
+    (uint32_t) now.tv_sec,
     now.tv_usec
   );
